Revert "Added solver for tetrominous"
This reverts commit 85182992a41791cba6df784fe751f162c9a96b27.
