+++ cspuz_rs_puzzles/src/puzzles/mod.rs
@@ -63,7 +63,7 @@ pub mod nurimaze;
 pub mod nurimisaki;
 pub mod parrot_loop;
 pub mod pencils;
-pub mod pentominous;
+pub mod polyominous;
 pub mod reflect;
 pub mod ringring;
 pub mod ripple;
@@ -87,7 +87,6 @@ pub mod statue_park;
 pub mod stostone;
 pub mod sudoku;
 pub mod tapa;
-pub mod tetrominous;
 pub mod the_longest;
 pub mod timebomb;
 pub mod tontonbeya;
