+++ cspuz_solver_backend/src/puzzle/mod.rs
@@ -132,7 +132,7 @@ puzzle_list!(puzz_link,
     (nurimaze, ["nurimaze"], "Nurimaze", "ぬりめいず"),
     (nurimisaki, ["nurimisaki"], "Nurimisaki", "ぬりみさき"),
     (pencils, ["pencils"], "Pencils", "ペンシルズ"),
-    (pentominous, ["pentominous"], "Pentominous", "Pentominous"),
+    (polyominous, ["pentominous"], "Pentominous", "Pentominous"),
     (reflect, ["reflect"], "Reflect Link", "リフレクトリンク"),
     (ringring, ["ringring"], "Ring-Ring", "リングリング"),
     (ripple, ["ripple"], "Ripple Effect", "波及効果"),
@@ -149,7 +149,6 @@ puzzle_list!(puzz_link,
     (stostone, ["stostone"], "Stostone", "ストストーン"),
     (sudoku, ["sudoku"], "Sudoku", "数独"),
     (tapa, ["tapa"], "Tapa", "Tapa"),
-    (tetrominous, ["tetrominous"], "Tetrominous", "Tetrominous"),
     (timebomb, ["timebomb"], "Time Bomb", "時限爆弾"),
     (tontonbeya, ["tontonbeya"], "Tontonbeya", "とんとんべや"),
     (yajikazu, ["yajikazu"], "Yajisan-Kazusan", "やじさんかずさん"),
